% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2dataframe.R
\name{fasta2dataframe}
\alias{fasta2dataframe}
\title{A Protein Sequence Function}
\usage{
fasta2dataframe(filepath)
}
\arguments{
\item{filepath}{path to FASTA file}
}
\value{
data.frame dataframe with a seq_name and sequence column
}
\description{
This function loads a FASTA file as a dataframe.
}
\examples{
fasta2dataframe("../go_sequences.fasta")
}
\keyword{data}
\keyword{fasta,}
\keyword{input}

