% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freqs.R
\name{get_freqs}
\alias{get_freqs}
\title{A Protein Sequence Function}
\usage{
get_freqs(x)
}
\arguments{
\item{x}{character vector containing permutations}
}
\value{
a list with components lengths and values
}
\description{
A Protein Sequence Function
}
\examples{
get_freqs(get_perms("ATAGACGATACGATACCCCGAGGGTAGGTA"))
}

