% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_vector.R
\name{to_vector}
\alias{to_vector}
\title{Converts sequences to numeric vector of primes products}
\usage{
to_vector(seq)
}
\arguments{
\item{seq}{protein sequence}
}
\value{
numeric vector of prime products
}
\description{
Converts sequences to numeric vector of primes products
}
\examples{
to_vector("ATAGACGATACGATACCCCGAGGGTAGGTA")
}

